Made re-usable concurrent create and apply
