+++ src/main/scala/com/github/plume/oss/passes/concurrent/PlumeCfgCreationPass.scala
@@ -2,7 +2,10 @@ package com.github.plume.oss.passes.concurrent
 
 import com.github.plume.oss.drivers.IDriver
 import com.github.plume.oss.passes.PlumeCpgPassBase
+import com.github.plume.oss.passes.concurrent.PlumeConcurrentCpgPass.producerQueueCapacity
+import com.github.plume.oss.passes.concurrent.PlumeConcurrentWriter.concurrentCreateApply
 import io.shiftleft.codepropertygraph.Cpg
+import io.shiftleft.codepropertygraph.generated.nodes.Method
 import io.shiftleft.passes.DiffGraph
 import io.shiftleft.semanticcpg.passes.controlflow.CfgCreationPass
 
@@ -23,54 +26,17 @@ class PlumeCfgCreationPass(cpg: Cpg) extends CfgCreationPass(cpg) with PlumeCpgP
 
   def createApplySerializeAndStore(driver: IDriver): Unit = {
     import PlumeCfgCreationPass.producerQueueCapacity
-    baseLogger.info(s"Start of enhancement: $name")
-    val nanosStart = System.nanoTime()
-    var nParts     = 0
-    var nDiff      = 0
-
-    init()
-    val parts = generateParts()
-    nParts = parts.length
-    val partIter        = parts.iterator
-    val completionQueue = mutable.ArrayDeque[Future[DiffGraph]]()
-    val writer          = new PlumeConcurrentWriter(driver, cpg)
-    val writerThread    = new Thread(writer)
-    writerThread.setName("Writer")
-    writerThread.start()
-    try {
-      try {
-        var done = false
-        while (!done) {
-          if (completionQueue.size < producerQueueCapacity && partIter.hasNext) {
-            val next = partIter.next()
-            completionQueue.append(Future.apply {
-              val builder = DiffGraph.newBuilder
-              runOnPart(builder, next)
-              builder.build()
-            })
-          } else if (completionQueue.nonEmpty) {
-            val future = completionQueue.removeHead()
-            val res    = Await.result(future, Duration.Inf)
-            nDiff += res.size
-            writer.queue.put(Some(res))
-          } else {
-            done = true
-          }
-        }
-      } finally {
-        try {
-          writer.queue.put(None)
-          writerThread.join()
-        } finally {
-          finish()
-        }
-      }
-    } finally {
-      val nanosStop = System.nanoTime()
-      baseLogger.info(
-        f"Enhancement $name completed in ${(nanosStop - nanosStart) * 1e-6}%.0f ms. ${nDiff}%d changes commited from ${nParts}%d parts."
-      )
-    }
+    concurrentCreateApply[Method](
+      producerQueueCapacity,
+      driver,
+      name,
+      baseLogger,
+      _ => init(),
+      _ => generateParts(),
+      cpg,
+      (x: DiffGraph.Builder, y: Method) => runOnPart(x, y),
+      _ => finish()
+    )
   }
 
 }
